# app/static/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Swagger API Materiales",
        "description": "You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.6"
    },
    "externalDocs": {
        "description": "Find out more about Swagger",
        "url": "http://swagger.io"
    }
    ,
    "tags": [
        {
            "name": "solicitudes",
            "description": "Solicitudes de materiales"
        },
        {
            "name": "materiales",
            "description": "Acerca de materiales"
        },
        {
            "name": "consultaFabricacion",
            "description": "Consultar estado de fabricación"
        }
    ],
    "paths": {
        "/materiales": {
            "get": {
                "tags": [
                    "materiales"
                ],
                "summary": "Listado de materiales",
                "description": "Obtiene listado de materiales disponibles junto a su ID",
                "responses": {
                    "200": {
                        "description": "successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Materiales"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sedes": {
            "get": {
                "tags": [
                    "sedes"
                ],
                "summary": "Listado de sedes",
                "description": "Obtiene listado de sedes disponibles junto a su ID y lo que fabrica",
                "responses": {
                    "200": {
                        "description": "successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sedes"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/pedidos/nuevo": {
            "post": {
                "tags": [
                    "pedidos"
                ],
                "summary": "Agregar solicitud",
                "requestBody": {
                    "description": "Parámetros para realizar la solicitud de materiales",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Solicitud"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Solicitud aceptada.",
                        "content": {}
                    },
                    "400": {
                        "description": "Solicitud inválida.",
                        "content": {}
                    },
                    "405": {
                        "description": "Método no permitido.",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "x-codegen-request-body-name": "body"
            }
        },
        "/pedidos/fabricacion": {
            "post": {
                "tags": [
                    "fabricacion"
                ],
                "summary": "Agregar solicitud de reserva de sede",
                "requestBody": {
                    "description": "Parámetros para realizar la reserva de sedes",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Fabricacion"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Solicitud aceptada.",
                        "content": {}
                    },
                    "400": {
                        "description": "Solicitud inválida.",
                        "content": {}
                    },
                    "405": {
                        "description": "Método no permitido.",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "x-codegen-request-body-name": "body"
            }
        },
        "/pedidos/consultarFabricacion": {
            "post": {
                "tags": [
                    "consultaFabricacion"
                ],
                "summary": "Consultar estado de la fabricación",
                "requestBody": {
                    "description": "Parámetros para realizar la consulta del estado de fabricación",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EstadoFabricacion"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Estado de la fabricación",
                        "content": {}
                    },
                    "404": {
                        "description": "No hay una reserva de fabricación con id",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "x-codegen-request-body-name": "body"
            }
        },
        "/pedidos/cancelar": {
            "post": {
                "tags": [
                    "Cancelar pedidos de materiales y sede"
                ],
                "summary": "Cancela pedidos de materiales y sede",
                "requestBody": {
                    "description": "Parámetros para realizar la consulta del estado de fabricación",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cancelar"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Estado de la fabricación",
                        "content": {}
                    },
                    "404": {
                        "description": "No hay pedido con ese id",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "x-codegen-request-body-name": "body"
            }
        }

    },
    "components": {
        "schemas": {
            "Materiales": {
                "type": "object",
                "properties": {
                    "id_material": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "nombre": {
                        "type": "string"
                    }
                }
            },
            "Sedes": {
                "type": "object",
                "properties": {
                    "sedes": {
                        "numero": {
                            "fabrica": {
                                "type": "string"
                            },
                            "id_sede": {
                                "type": "integer",
                                "format": "int32"
                            },
                            "nombre":{
                                "type": "string"
                            }
                        }
                    },
                    "nombre": {
                        "type": "string"
                    }
                }
            },
            "Solicitud": {
                "type": "object",
                "properties": {
                    "fecha_necesaria_entrega": {
                        "type": "string",
                        "format": "date"
                    },
                    "materiales": {
                        "type": "object",
                        "example": "{\"4\":\"20\",\"3\":\"13\"}"
                    }
                }
            },
            "EstadoFabricacion": {
                "type": "object",
                "properties": {
                    "idPedidoSede": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "Cancelar": {
                "type": "object",
                "properties": {
                    "idPedido": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "idPedidoSede": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "Fabricacion": {
                "type": "object",
                "properties": {
                    "fecha_reserva_sede": {
                        "type": "string",
                        "format": "date"
                    },
                    "fecha_necesaria_entrega": {
                        "type": "string",
                        "format": "date"
                    },
                    "pedido_fabricacion": {
                        "type": "object",
                        "example": "{\"materialMarcos\":\"1\",\"cantidadMarcos\":\"250\",\"materialPatillas\":\"1\",\"cantidadPatillas\":\"200\",\"materialEstuche\":\"2\",\"cantidadEstuche\":\"3\",\"materialLentes\":\"3\",\"cantidadLentes\":\"290\"}"
                    }
                }
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}